feat:resend otp for email verification
feat: Add OTP resend functionality for email verification

This change introduces a new endpoint and controller function to resend
the OTP (One-Time Password) for email verification. The
`resendOtpController` has been implemented in the `auth.controller.js`
file, along with necessary modifications in the `auth.services.js` file
to handle OTP generation and update user data.

The purpose of this feature is to allow users who have not yet verified
their email addresses to request a new OTP in case the previous one has
expired or been lost. When invoked, the controller verifies the email
address, generates a new OTP, updates the user's OTP data, and sends the
new OTP to the user's email address for verification.

This commit adds support for resending OTPs, thereby enhancing the user
experience and ensuring a more seamless email verification process.