refactor response handling in auth controller and error handling in error handler
The changes in this commit refactor the response handling in the
authentication controller and the error handling in the error handler
middleware. Instead of declaring response variables outside the if-else
blocks, they are now declared within the corresponding blocks to ensure
proper scoping. Additionally, a handleErrors function is introduced in
the auth services to centralize error handling and replace repetitive
code. This makes the code more concise and easier to maintain.
the bug related to this commit is no response body in reset endpoint
because 204 status code should has no content

Refactor response and error handling in auth controller and error
handler

This commit refactors the response handling in the authentication
controller by declaring response variables within the corresponding
if-else blocks for proper scoping. It also introduces a handleErrors
function in the auth services to centralize error handling and eliminate
repetitive code. The changes improve code conciseness and
maintainability. Additionally, this commit fixes a bug related to the
reset endpoint where no response body was sent to client so to ensure
that body sent the status code changed to 200